// demo-ui/src/app/features/models/reference/localization/views/basic/CityBasicView.ts
/**
 * auto generated basic view bean ts file
 * <br/>write modifications between specific code marks
 * <br/>processed by skeleton-generator
 */
export class CityBasicView {

id: number;
selected: boolean;
canDelete: boolean;
regionCountryCode: string;
regionCode: string;
code: string;
label: string;

/* Specific Code Start */
/* Specific Code End */
}