// demo-ui/src/app/features/models/reference/localization/sortings/CitySorting.ts
/**
 * auto generated sorting ts file
 * <br/>write modifications between specific code marks
 * <br/>processed by skeleton-generator
 */
export class CitySorting {

regionCountryCodeOrderType: string;
regionCodeOrderType: string;
codeOrderType: string;
labelOrderType: string;

/* Specific Code Start */
/* Specific Code End */
}